Refactors scan results list component
