+++ dashboard/src/components/pages/Header.js
@@ -5,7 +5,9 @@ export default function Header() {
   return (
     <nav className="flex items-center justify-between flex-wrap bg-teal-500 p-6">
       <div className="flex items-center flex-shrink-0 text-white mr-6">
-        <span className="font-semibold text-xl tracking-tight">GuardRails Challenge</span>
+        <Link to="/" className="font-semibold text-xl tracking-tight">
+          GuardRails Challenge
+        </Link>
       </div>
 
       <div className="block lg:hidden">
